Extracting out connection service
