+++ src/app/services/connection.service.ts
@@ -1,0 +1,291 @@
+import { Injectable } from '@angular/core';
+import { NeonGTDConfig } from '../neon-gtd-config';
+import { query } from 'neon-framework';
+
+export interface RequestWrapper {
+    abort(): void;
+    always(callback: Function): void;
+    done(callback: Function): void;
+    fail(callback: Function): void;
+}
+
+export interface Connection<T extends { query: any } = { query: any }> {
+
+    /**
+     * Deletes the saved dashboard state with the given name.
+     *
+     * @arg {string} stateName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    deleteState(stateName: string, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Returns the accessible database names.
+     *
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    getDatabaseNames(onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Returns the types of the fields in the given database/table.
+     *
+     * @arg {string} databaseName
+     * @arg {string} tableName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    getFieldTypes(databaseName: string, tableName: string, onSuccess: (response: any) => void,
+        onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Returns the saved dashboard states.
+     *
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    listStates(limit: number, offset: number, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Returns the table and field names in the given database.
+     *
+     * @arg {string} databaseName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    getTableNamesAndFieldNames(databaseName: string, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Loads the saved state with the given name.
+     *
+     * @arg {string} stateName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    loadState(stateName: string, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Runs an export query with the given data and format.
+     *
+     * @arg {any} exportData
+     * @arg {any} exportFormat
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    runExportQuery(exportData: any, exportFormat: any, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Runs a search query with the given payload.
+     *
+     * @arg {T} queryPayload
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    runSearchQuery(queryPayload: T, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+
+    /**
+     * Saves (or overwrites) a state with the given data.
+     *
+     * @arg {{dashboards:Dashboard,datastores:Datastore[],layouts:any,stateName:string}} stateData
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @abstract
+     */
+    saveState(stateData: NeonGTDConfig, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper;
+}
+
+// Internal class that wraps AbstractSearchService.Connection.  Exported to use in the unit tests.
+export class NeonConnection<T extends { query: any } = { query: any }> implements Connection<T> {
+    constructor(public connection: query.Connection) { }
+
+    /**
+     * Deletes the saved dashboard state with the given name.
+     *
+     * @arg {string} stateName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public deleteState(stateName: string, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper {
+        return this.connection.deleteState(stateName, onSuccess, onError);
+    }
+
+    /**
+     * Returns the accessible database names.
+     *
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public getDatabaseNames(onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper {
+        return this.connection.getDatabaseNames(onSuccess, onError);
+    }
+
+    /**
+     * Returns the types of the fields in the given database/table.
+     *
+     * @arg {string} databaseName
+     * @arg {string} tableName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public getFieldTypes(
+        databaseName: string,
+        tableName: string,
+        onSuccess: (response: any) => void,
+        onError?: (response: any) => void
+    ): RequestWrapper {
+        return this.connection.getFieldTypes(databaseName, tableName, onSuccess, onError);
+    }
+
+    /**
+     * Returns the saved dashboard states.
+     *
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public listStates(limit: number, offset: number, onSuccess: (response: any) => void,
+        onError?: (response: any) => void): RequestWrapper {
+        return this.connection.listStates(limit, offset, onSuccess, onError);
+    }
+
+    /**
+     * Returns the table and field names in the given database.
+     *
+     * @arg {string} databaseName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public getTableNamesAndFieldNames(
+        databaseName: string,
+        onSuccess: (response: any) => void,
+        onError?: (response: any) => void
+    ): RequestWrapper {
+        return this.connection.getTableNamesAndFieldNames(databaseName, onSuccess, onError);
+    }
+
+    /**
+     * Loads the saved state with the given name.
+     *
+     * @arg {string} stateName
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public loadState(stateName: string, onSuccess: (response: any) => void, onError?: (response: any) => void): RequestWrapper {
+        return this.connection.loadState({
+            stateName: stateName
+        }, onSuccess, onError);
+    }
+
+    /**
+     * Runs an export query with the given data and format.
+     *
+     * @arg {any} exportData
+     * @arg {any} exportFormat
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public runExportQuery(
+        exportData: { data: any },
+        exportFormat: any,
+        onSuccess: (response: any) => void,
+        onError?: (response: any) => void
+    ): RequestWrapper {
+        return this.connection.executeExport(exportData, onSuccess, onError, exportFormat);
+    }
+
+    /**
+     * Runs a search query with the given payload.
+     *
+     * @arg {T} queryPayload
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public runSearchQuery(
+        queryPayload: T,
+        __onSuccess: (response: any) => void,
+        __onError?: (response: any) => void
+    ): RequestWrapper {
+        return this.connection.executeQuery(queryPayload.query, null);
+    }
+
+    /**
+     * Saves (or overwrites) a state with the given data.
+     *
+     * @arg {{dashboards:Dashboard,datastores:Datastore[],layouts:any,stateName:string}} stateData
+     * @arg {(response: any) => void} onSuccess
+     * @arg {(response: any) => void} [onError]
+     * @return {RequestWrapper}
+     * @override
+     */
+    public saveState(
+        stateData: NeonGTDConfig,
+        onSuccess: (response: any) => void,
+        onError?: (response: any) => void
+    ): RequestWrapper {
+        return this.connection.saveState(stateData, onSuccess, onError);
+    }
+}
+
+@Injectable({
+    providedIn: 'root'
+})
+export class ConnectionService {
+    // Maps the datastore types to datastore hosts to connections.
+    private connections: Map<string, Map<string, NeonConnection<any>>> = new Map<string, Map<string, NeonConnection<any>>>();
+
+    /**
+     * Returns an existing connection to the REST server using the given host and the given datastore type (like elasticsearch or sql), or
+     * creates and returns a Neon connection if none exists.
+     */
+    public connect<T extends { query: any } = { query: any }>(datastoreType: string, datastoreHost: string): NeonConnection<T> {
+        if (datastoreType && datastoreHost) {
+            if (!this.connections.has(datastoreType)) {
+                this.connections.set(datastoreType, new Map<string, NeonConnection<T>>());
+            }
+            if (!this.connections.get(datastoreType).has(datastoreHost)) {
+                let connection = this.neonConnection();
+                connection.connect(datastoreType, datastoreHost);
+                this.connections.get(datastoreType).set(datastoreHost, new NeonConnection<T>(connection));
+            }
+            return this.connections.get(datastoreType).get(datastoreHost);
+        }
+        return null;
+    }
+
+    private neonConnection(): query.Connection {
+        return new query.Connection();
+    }
+}
